Renamed to skeleton
