+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/templates/commandbasedskeleton/RobotContainer.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpilibj.templates.barebonescommandbased;
+package edu.wpi.first.wpilibj.templates.commandbasedskeleton;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
